+++ src/config/sitesettings.ts
@@ -1,0 +1,7 @@
+﻿/**
+ * Site Settings Configuration Re-export Module
+ * Shri Trinetra Mahakal Mandir & Registered Dharmic Trust
+ */
+
+export * from './site';
+export { SITE_CONFIG as default } from './site';